requirements.txt
streamlit
sqlalchemy
python-dotenv
pydantic
pandas
